libv4l: add ability to determine flags based on DMI info
From: Hans de Goede <hdegoede@redhat.com>

It is possible for the same laptop webcam module (so same usb id) to
be mounted upside down in some models and the right way up in
other laptop models. This patch adds the ability to only apply
flags to a webcam based on the combination of usb id and
dmi info to identify the laptop model.

It also adds the webcam in the Asus N50Vn as the first upside down cam
identified this way.

Priority: normal

Signed-off-by: Hans de Goede <hdegoede@redhat.com>