libv4l: add software autogain / exposure
From: Hans de Goede <hdegoede@redhat.com>

Add software autogain / exposure, for camera's which have gain and
exposure controls but do not contain the ability to calculate the average
lumination in hardware (which is needed to do this in the kernel). This
patch enables this for the spca561 rev12a, but it should be usefull for
other cameras too.

Priority: normal

Signed-off-by: Hans de Goede <hdegoede@redhat.com>