libv4l: Fix reqbuf Device or Resource busy error when using read
From: Hans de Goede <hdegoede@redhat.com>

Some applications such as v4l2-apps/test/capture-example.c, in read mode
use select() together with read() and do a select() before the first read().

This causes issues together with certain drivers (gspca for example),
do not allow switching from read mode to mmap mode and they assume read()
mode if a select or poll() is done before any buffers are requested.

When not using libv4l2, this is not an issue but libv4l2 uses mmap mode
under the hood when converting as that safes a memcpy for each frame read.

This fails with such drivers when the application has done a select() before
the first read() as the driver now is in "read mode" and disallows switching
to mmap mode.

This patch fixes this by falling back to using read() for v4l2_read() when
using mmap mode fails.

Priority: normal

Signed-off-by: Hans de Goede <hdegoede@redhat.com>